requirements.txt
fastapi
uvicorn[standard]
requests
python.env
pydantic
python-dotenv
PyMuPDF
nltk
tiktoken
unstructured
openai
tenacity
numpy
pinecone-client
google-cloud-storage
google-cloud-pubsub
google-cloud-functions
google-cloud-logging
google-cloud-core
orjson
black
ruff
pylint
mypy
ipython
pytest
httpx